refactor: 공유하기 features로 분리 (#51)
