+++ src/features/share/index.ts
@@ -1,0 +1,1 @@
+export { default as ShareButton } from './ui/ShareButton'
