+++ src/widgets/playlist/ActionBar.tsx
@@ -3,9 +3,9 @@ import { useNavigate } from 'react-router-dom'
 
 import styled, { useTheme } from 'styled-components'
 
-import { Heart, Share, Playlist } from '@/assets/icons'
-import { flexColCenter, flexRowCenter } from '@/shared/styles/mixins'
-import { BottomSheet, Button } from '@/shared/ui'
+import { Heart, Playlist } from '@/assets/icons'
+import { ShareButton } from '@/features/share'
+import { flexRowCenter } from '@/shared/styles/mixins'
 import SvgButton from '@/shared/ui/SvgButton'
 
 interface ActionBarProps {
@@ -15,27 +15,14 @@ interface ActionBarProps {
 const ActionBar = ({ playlistId }: ActionBarProps) => {
   const theme = useTheme()
   const [liked, setLiked] = useState(false)
-  const [isBottomSheetOpen, setIsBottomSheetOpen] = useState(false)
   const navigate = useNavigate()
 
   const handleLike = () => {
     setLiked((prev) => !prev)
   }
 
-  const handleShare = () => {
-    setIsBottomSheetOpen(true)
-  }
-
   const handleMovePlaylist = () => {
     navigate(`/discover/${playlistId}/playlist`)
-  }
-
-  // TODO: 이미지 저장 기능 구현
-  const handleSaveImage = () => {}
-
-  const handleCopyLink = () => {
-    const link = `${window.location.origin}/discover/${playlistId}`
-    navigator.clipboard.writeText(link)
   }
 
   return (
@@ -48,26 +35,8 @@ const ActionBar = ({ playlistId }: ActionBarProps) => {
         stroke={liked ? theme.COLOR['primary-normal'] : theme.COLOR['gray-200']}
         onClick={handleLike}
       />
-      <SvgButton icon={Share} width={24} height={24} onClick={handleShare} />
+      <ShareButton playlistId={playlistId} />
       <SvgButton icon={Playlist} width={24} height={24} onClick={handleMovePlaylist} />
-
-      <BottomSheet
-        isOpen={isBottomSheetOpen}
-        onClose={() => setIsBottomSheetOpen(false)}
-        height="fit-content"
-      >
-        <BottomSheetWrapper>
-          <ImagePreview />
-          <ButtonBar>
-            <Button onClick={handleSaveImage} size="M" state="secondary">
-              이미지로 저장
-            </Button>
-            <Button onClick={handleCopyLink} size="M" state="secondary">
-              링크 복사
-            </Button>
-          </ButtonBar>
-        </BottomSheetWrapper>
-      </BottomSheet>
     </Wrapper>
   )
 }
@@ -77,26 +46,4 @@ export default ActionBar
 const Wrapper = styled.div`
   ${flexRowCenter}
   gap: 16px;
-`
-
-const ImagePreview = styled.div`
-  width: 280px;
-  height: 280px;
-  border-radius: 24px;
-  background-color: ${({ theme }) => theme.COLOR['gray-900']};
-`
-
-const BottomSheetWrapper = styled.div`
-  ${flexColCenter}
-  gap: 24px;
-`
-
-const ButtonBar = styled.div`
-  ${flexRowCenter}
-  gap: 10px;
-  width: 100%;
-
-  & button {
-    flex: 1;
-  }
 `
